requirements.txt
python-telegram-bot==12.7
PyMySQL
mysql-connector-python
requests
openai
deezepy
stripe
flask
schedule